Move builder inside of producer, remove unnecessary perf tests
