--- src/test/resources/decrypted.properties
@@ -1,2 +1,0 @@
-secretKey=PamXL2ltvVszqOw/GdC+4lh20ELAV00BN/IFhNbW
-accessKey=AKIAJOVDQEGNMT5APG7Q
