--- src/test/resources/example.config.properties
@@ -1,15 +1,0 @@
-user.name=YOUR_USERNAME
-user.password=YOUR_PASSWORD
-account.name=YOUR_ACCOUNT_NAME
-product=track
-stream.label=YOUR_STREAM_LABEL
-
-accessKey=YOUR_AWS_ACCESS_KEY
-secretKey=YOUR_AWS_SECRET_KEY
-
-kinesis.stream.name=YOUR_AWS_KINESIS_STREAM_NAME
-
-thread.count=10
-batch.size=50000
-message.queue.size=300
-bytes.queue.size=300
