--- src/test/resources/test.properties
@@ -1,3 +1,0 @@
-msgSize=30000
-maxThreadCount=20
-messagesPerThread=100
